Allow passing an optional ExecutorPlugin to BoostExecutor.
This allows someone with a file-descriptor based interface to select on
their file descriptor in their main dispatcher, and make this plugin set
their file descriptor dirty (write to it) whenever the main dispatcher
needs to be woken up.

Caveat: This does not help with remote connections at all, since this
plugin cannot do anything when there's data on io_service via the boost
socket.

Closes #60.